SSCSCI-969: use latest validator from common
