+++ build.gradle
@@ -351,8 +351,7 @@ dependencies {
     implementation group: 'org.flywaydb', name: 'flyway-core', version: '9.22.3'
     implementation group: 'org.postgresql', name: 'postgresql', version: '42.6.0'
 
-    implementation group: 'com.github.hmcts', name: 'sscs-common', version: '5.12.8'
-
+    implementation group: 'com.github.hmcts', name: 'sscs-common', version: '9671137'
     implementation group: 'com.github.hmcts', name: 'sscs-pdf-email-common', version: '5.2.14'
 
     implementation group: 'com.azure', name: 'azure-core', version: '1.46.0'
